% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_1440.R
\name{bb_1440}
\alias{bb_1440}
\title{Summarize Biobank Acceleration File into 1440 format}
\usage{
bb_1440(df, summarize_over_day = FALSE, summarize_func = "mean",
  na.rm = TRUE, ...)
}
\arguments{
\item{df}{A \code{data.frame} or \code{tbl} from \code{\link{bb_read}}}

\item{...}{Additional arguments to pass to \code{\link{bb_summarize}}}
}
\value{
A table of the day (if \code{summarize_over_day = FALSE}) and 1440 columns
}
\description{
Summarize Biobank Acceleration File into 1440 format
}
