% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_read.R
\name{bb_read}
\alias{bb_read}
\title{Read Biobank Acceleration File}
\usage{
bb_read(file, force_numeric = TRUE, tz = "UTC", max_days = 7)
}
\arguments{
\item{file}{path to filename}

\item{force_numeric}{Should the acceleration be
set to numeric even if read in as character
(odd but happens)}

\item{tz}{timezone passed to \code{lubridate} functions}

\item{max_days}{Maximum number of days the data should have.  Has no effect
on reading in the data, but prints a warning if there are more than the
maximum.}
}
\value{
A \code{tibble}
}
\description{
Read Biobank Acceleration File
}
\examples{
file = system.file("test2.csv", package = "biobankr")
  df = bb_read(file)
}
