% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_read_html.R
\name{bb_read_html}
\alias{bb_read_html}
\title{Read Biobank Document HTML}
\usage{
bb_read_html(file)
}
\arguments{
\item{file}{Name of HTML file}
}
\value{
A list of `data.frames` of the tables
}
\description{
Read Biobank Document HTML
}
