% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_summarize.R
\name{bb_summarize}
\alias{bb_summarize}
\alias{bb_summarise}
\alias{bb_summarize_minute}
\title{Summarize Biobank Acceleration File}
\usage{
bb_summarize(df, unit = "1 minute", summarize_func = "mean",
  na.rm = TRUE, keep_imputed = TRUE, ...)

bb_summarise(df, unit = "1 minute", summarize_func = "mean",
  na.rm = TRUE, keep_imputed = TRUE, ...)

bb_summarize_minute(..., unit = "1 minute")
}
\arguments{
\item{df}{A \code{data.frame} or \code{tbl} from \code{\link{bb_read}}}

\item{unit}{unit of summarization, passed to
\code{\link[lubridate]{floor_date}}}

\item{summarize_func}{Function to use for summarization,
passed to \code{\link{summarize}}}

\item{na.rm}{Should NAs be removed?}

\item{keep_imputed}{Should the imputed values be kept in this data set
before summarization}

\item{...}{Additional arguments to pass to \code{summarize_func}}
}
\value{
A summarized \code{tbl}
}
\description{
Summarize Biobank Acceleration File
}
\examples{
file = system.file("test2.csv", package = "biobankr")
df = bb_read(file)
collapsed = bb_summarize(df, unit = "10 seconds")
collapsed = bb_summarize(df, unit = "10 seconds", summarize_func = "median")
collapsed = bb_summarize(df, unit = "15 seconds",
summarize_func = "quantile", probs = 0.25)
}
