% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_day_min.R
\name{date_day_min}
\alias{date_day_min}
\title{Convert Date to Day/Minute}
\usage{
date_day_min(df, from_baseline = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} or \code{tbl} from \code{\link{bb_read}}}

\item{from_baseline}{Should day be day from baseline or
the day in Date format}
}
\value{
A summarized \code{tbl}
}
\description{
Convert Date to Day/Minute
}
\examples{
file = system.file("test2.csv", package = "biobankr")
df = bb_read(file)
day_stuff = date_day_min(df)
date_stuff = date_day_min(df, from_baseline = FALSE)
}
