% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_info.R
\name{extract_sampling_rate}
\alias{extract_sampling_rate}
\alias{extract_date}
\title{Extract information from header}
\usage{
extract_sampling_rate(info)

extract_date(info)
}
\arguments{
\item{info}{Header information from csv header}
}
\value{
Sampling rates or dates
}
\description{
Extract information from header
}
\examples{
file = system.file("test2.csv", package = "biobankr")
  df = readr::read_csv(file = file)
 info = colnames(df)[1]
 info = trimws(info)
 srate = extract_sampling_rate(info)
date_range = extract_date(info)
}
