% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cwa.R
\name{read_cwa}
\alias{read_cwa}
\title{Read CWA from Axtivity}
\usage{
read_cwa(file, end = Inf, verbose = TRUE, convert_time = TRUE)
}
\arguments{
\item{file}{filename of cwa file}

\item{end}{End point for reading data.  Default is all the data}

\item{verbose}{print diagnostic messages}

\item{convert_time}{Should times be converted to \code{POSIXct}?}
}
\value{
A list with header information and a tbl
}
\description{
Read CWA from Axtivity
}
