% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_start_stop.R
\name{test_start_stop}
\alias{test_start_stop}
\title{Testing if there are too many cases}
\usage{
test_start_stop(df, start_hour = 10, end_hour = 10)
}
\arguments{
\item{df}{The \code{data.frame} output from \code{\link{bb_read}}.  Should have
a \code{too_many_days} attribute}

\item{start_hour}{Hour to check if the data has started on this hour}

\item{end_hour}{Hour to check if the data has ended on this hour}
}
\value{
A filtered \code{data.frame}
}
\description{
Testing if there are too many cases
}
