% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb_data.R
\name{ukb_data}
\alias{ukb_data}
\title{Extract UKB Data}
\usage{
ukb_data(fileset, path = ".")
}
\arguments{
\item{fileset}{The prefix for a UKB fileset, e.g., \code{ukbxxxx}}

\item{path}{The path to the directory containing your UKB fileset.
The default value is the current directory.}
}
\value{
A list of the data and a lookup table
}
\description{
Extract UKB Data
}
\note{
This function is a light wrapper of \code{\link{ukb_df}} and
\code{\link{ukb_df_field}} but does things using a temporary
directory.
}
