% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_1440.R
\name{bb_1440}
\alias{bb_1440}
\title{Summarize Biobank Acceleration File}
\usage{
bb_1440(df, ...)
}
\arguments{
\item{df}{A \code{data.frame} or \code{tbl} from \code{\link{bb_read}}}

\item{...}{Additional arguments to pass to \code{\link{bb_summarize}}}
}
\value{
A table of the day and 1440 columns
}
\description{
Summarize Biobank Acceleration File
}
