% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_1440.R
\name{bb_1440}
\alias{bb_1440}
\title{Summarize Biobank Acceleration File}
\usage{
bb_1440(df, ...)
}
\arguments{
\item{df}{A \code{data.frame} or \code{tbl} from \code{\link{bb_read}}}

\item{...}{Additional arguments to pass to \code{summarize_func}}

\item{unit}{unit of summarization, passed to
\code{\link[lubridate]{floor_date}}}

\item{summarize_func}{Function to use for summarization,
passed to \code{\link{summarize}}}

\item{keep_imputed}{Should the imputed values be kept in this data set
before summarization}
}
\value{
A summarized \code{tbl}
}
\description{
Summarize Biobank Acceleration File
}
