% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cwa.R
\name{read_cwa}
\alias{read_cwa}
\title{Read CWA from Axtivity}
\usage{
read_cwa(file, end = Inf, convert_time = TRUE, verbose = TRUE,
  tz = "", ...)
}
\arguments{
\item{file}{filename of cwa file}

\item{end}{End point for reading data.  Default is all the data}

\item{convert_time}{Should times be converted to \code{POSIXct}?}

\item{verbose}{print diagnostic messages}

\item{tz}{A time zone specification to be used for the conversion,
if one is required.  System-specific (see \link{timezones}),
but "" is the current time zone,
and "GMT" is UTC (Universal Time, Coordinated).}

\item{...}{additional arguments to pass to \code{\link{g.cwaread}},
noting some are already loaded in different arguments}
}
\value{
A list with header information and a tbl
}
\description{
Read CWA from Axtivity
}
