% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_helpers.R
\name{time_to_min}
\alias{time_to_min}
\alias{min_to_time}
\title{Time helper functions}
\usage{
time_to_min(x)

min_to_time(x, day = NULL)
}
\arguments{
\item{x}{an object of class Date or numeric time.  Note,
only seconds are kept, not milliseconds}

\item{day}{A date given in YMD format}
}
\value{
A Date or numeric object
}
\description{
Time helper functions
}
\examples{
tm = Sys.time()
print(tm)
mins = time_to_min(tm)
mins
min_to_time(mins)
min_to_time(mins, day = "2017-01-24")
}
